[paywall] Show each custom claim as its own column
