+++ web/store/aws/awsSelectors.ts
@@ -1,8 +1,9 @@
 import { createSelector } from 'reselect'
 
 import { fromCognitoIdentityPool } from '@aws-sdk/credential-providers'
+import { LambdaClient } from '@aws-sdk/client-lambda'
 
-import { clientConfig, identityPoolId } from '../aws'
+import { clientConfig, identityPoolId } from './aws'
 import { firebaseLogins } from 'store/auth/firebase'
 import { getFirebaseToken } from 'store/auth/authSelectors'
 
@@ -15,5 +16,13 @@ export const getAwsCredentials = createSelector(
         logins: firebaseLogins(firebaseToken),
         clientConfig,
       })
+  }
+)
+
+export const getLambdaClient = createSelector(
+  [getAwsCredentials],
+  credentials => {
+    if (credentials)
+      return new LambdaClient({ region: 'us-east-2', credentials })
   }
 )
